+++ client/src/app/components/adminCC-dashboard/medication-listing/medication-listing.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute, NavigationEnd } from '@angular/router';
 
 import { ClinicalCenterService } from 'src/app/services/clinical-center.service';
+import { DialogService } from 'src/app/services/dialog.service';
 
 @Component({
 	selector: 'app-medication-listing',
@@ -16,7 +17,8 @@ export class MedicationListingComponent implements OnInit {
 
 	constructor(private clinicalCenterService: ClinicalCenterService,
 		private router: Router,
-		private activatedRoute: ActivatedRoute) {
+		private activatedRoute: ActivatedRoute,
+		private confirmationDialogService: DialogService) {
 
 		this.navigationSubscription = this.router.events.subscribe((e: any) => {
 			if (e instanceof NavigationEnd) {
@@ -41,10 +43,15 @@ export class MedicationListingComponent implements OnInit {
 	}
 
 	deleteMedication(medication: any) {
-		this.clinicalCenterService.deleteMedication(medication.id).subscribe(
-			(data) => this.getMedications(),
-			(error) => alert(error)
-		);
+		this.confirmationDialogService.confirm('Please confirm', 'Are you sure you want to delete medication with code: ' + medication.code + ' ?', false)
+		.then((confirmed) => {
+			if (confirmed.submited) {
+				this.clinicalCenterService.deleteMedication(medication.id).subscribe(
+					(data) => this.getMedications(),
+					(error) => alert(error)
+				);
+			}
+		});
 	}
 
 	ngOnDestroy() {
